requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
pycryptodome==3.20.0
requests==2.31.0
requests-toolbelt==1.0.0
urllib3==2.2.0
